very good
This is very good.  Only critique is that some of your style feels like it came from a different language - and is not rubyish.  For example, to call a ruby method without an argument, we don't need the the (), and ruby has an implicit return at the end of a method, so don't need to write the return statement.